Use example.com in example email addresses and default values.
There are a number of domains specifically meant for examples
and similar bogus values, so let's use them instead of risking
being a disturbance for the people who own the real domains used
in our examples and default values.

We change the default value of the From address (the 'from' option)
which theoretically could cause problems for users who somehow
have managed to rely on the default being logstash.alert@nowhere.com.